requirements.txt
streamlit
pylast
google-genai
plotly
pandas